+++ NEWS.md
@@ -3,17 +3,35 @@
 * `get_correlation()` now respects `dplyr::group_by()` grouping
 
 
-# moderndive 0.2.0.9000
 
-* Added ID argument to get_regression_points() to return a column that ID's the 
-observational units
-* Removed all `assertive::assert()`
+
+***
+
+
+
+# moderndive 0.3.0
+
+* Added minimally viable "parallel slopes" regression model plotting function `gg_parallel_slopes()`. In the future we hope to define a new `ggplot2` geom.
+* Added "Why `moderndive`?" vignette
+* Added ID argument to `get_regression_points()` to return a column that identifies the 
+observational units/rows
+* Datasets:
+        + Added `DD_vs_SB`: Dunkin Donuts and Starbucks in Eastern Massachusetts data collected by @DelaneyMoran
+        + Added  `promotions`: tibble version of `openintro::gender.discrimination` used to illustrate permutation test.
+        + Added `MA_schools`: Relationship between SAT scores and socio-economic status for Massachusetts high schools.
+        + Added `mythbusters_yawn`: Data from study on Mythbusters show on whether yawning is
+        + Added  `promotions_shuffled`: one instance of `promotions` with `gender` permuted/shuffled
+        + Original `pennies_sample` sample of 40 pennies from `pennies` has been renamed `orig_pennies_sample`. New `pennies_sample` consists of 50 pennies sampled from bank in Northampton, MA, USA on 2019/2/1.
+        + Added  `pennies_resamples`: 35 bootstrap resamples of new `pennies_sample`
+        + Added  `movies_genre`: random sample of 32 action and 36 romance movies from `ggplot2movies::movies`        
+* Removed all `assertive::assert()` code
 * Converted `house_prices$date` from `dttm` (date-time) to `date` per R4DS [comment](https://r4ds.had.co.nz/dates-and-times.html#creating-datetimes) on using simplest data type possible
-* Added `DD_vs_SB` dataset of Dunkin Donuts and Starbucks in Eastern Massachusetts collected by @DelaneyMoran
-* Updated `get_correlation()` to:
-    + Work with `group_by()`
-    + Handle missing values
-    + Handle more than two variables
+
+
+
+***
+
+
 
 # moderndive 0.2.0
 
@@ -33,9 +51,17 @@ Details:
     - Otherwise omit `residual`
 * Removed `tidyverse` from Depends, Imports, or Suggests
 
+
+***
+
+
 # moderndive 0.1.1
 
 Fixed broken url in `?bowl_samples`
+
+
+***
+
 
 # moderndive 0.1.0